+++ src/lib.js
@@ -1,6 +1,8 @@
 const path = require('path');
 const fs = require('fs');
 const { promisify } = require('util');
+const SHELL_NODE_MODULES_PATH = process.env.SHELL_NODE_MODULES_PATH;
+const webpack = require(path.join(SHELL_NODE_MODULES_PATH, 'webpack'));
 
 const PROJECT_ROOT = exports.PROJECT_ROOT = process.cwd();
 const SOURCE_DIR = exports.SOURCE_DIR = path.join(PROJECT_ROOT, 'src');
@@ -27,3 +29,35 @@ exports.getFilesByExtName = async (pathname, extname) => {
 
   return fileList;
 };
+
+exports.getHtmlLoaderConfig = PROJECT_CONFIG => ({
+  test: /\.html?$/,
+  use: [
+    {
+      loader: require.resolve('html-loader'),
+      options: {
+        interpolate: true,
+        root: './',
+      },
+    },
+    {
+      loader: require.resolve('posthtml-loader'),
+    },
+  ],
+});
+
+exports.getLoaderOptionPlugin = PROJECT_CONFIG => new webpack.LoaderOptionsPlugin({
+  options: {
+    posthtml(ctx) {
+      return {
+        plugins: [
+          require('posthtml-expressions')({
+            locals:{
+              env: process.env.NODE_ENV, 
+            }
+          })
+        ],
+      };
+    }
+  },
+});