+++ test/index.test.js
@@ -42,7 +42,7 @@ describe('exports methods: <<getProdConfig>> and <<getDevConfig>>', () => {
     );
     expect(devConfig.module.strictExportPresence).to.be.true;
     expect(devConfig.module.rules[0].oneOf).to.have.lengthOf(6);
-    expect(devConfig.plugins).to.have.lengthOf(5);
+    expect(devConfig.plugins).to.have.lengthOf(6);
   });
 
   it('should getProdConfig a valid webpack config', async () => {
@@ -64,7 +64,7 @@ describe('exports methods: <<getProdConfig>> and <<getDevConfig>>', () => {
     );
     expect(prodConfig.module.strictExportPresence).to.be.true;
     expect(prodConfig.module.rules[0].oneOf).to.have.lengthOf(6);
-    expect(prodConfig.plugins).to.have.lengthOf(7);
+    expect(prodConfig.plugins).to.have.lengthOf(8);
   });
 
 });